requirements.txt
flask
gunicorn
tensorflow-cpu
numpy